Added notes throughout
I've reached a good stopping point now.

Getting a bit fed up with this project.

I'm not at all happy with the design but I could
obviously change this - the functionality is more
important.

The only bits of functionality that I want/need to
add are:
 - Ability to move a task to a different list
 - The local storage so you can save to do lists
 - Do something with the dates - and make them
 changeable

I will do all of these in the future but for now
I'm going to move on to the next thing in TOP.